move pixelator.mpx.plot (REMEMBER: update tutorial import statement)
