+++ src/pixelator/plot/constants.py
