+++ tests/plot/__init__.py
@@ -1,0 +1,4 @@
+"""Top-level plotting helpers for Pixelator.
+
+Copyright © 2023 Pixelgen Technologies AB.
+"""
