+++ tests/plot/data/density_scatter_plot_input.h5ad
