+++ tests/plot/snapshots/test_density_scatter_plot/test_density_scatter_plot_CD3-CD8-extra_params1.png
